Fix types

+++ app/database/methods.go
@@ -1,8 +1,8 @@
 package database
 
-// GetTotal returns the total amount of a payment order
-func (order PaymentOrder)GetTotal() (amount float32) {
-	amount = 0.0
+// GetTotal returns the total amount of a payment order in cents
+func (order PaymentOrder) GetTotal() (amount int) {
+	amount = 0
 	for _, item := range order.OrderItems {
 		amount += item.Price * item.Quantity
 	}
